Split out `sortItemsByDate`
